moved bulk of work to build, run simply launches the built bits, removed zipping
